+++ extended-ref/gasnet_coll_scratch.c
@@ -408,7 +408,7 @@ int8_t gasnete_coll_scratch_alloc_nb(gasnete_coll_op_t* op GASNETE_THREAD_FARG) 
   if(scratch_req->incoming_size > scratch_req->team->scratch_segs[scratch_req->team->myrank].size) {
     gasneti_fatalerror("%d> collective requires temporary storage (%ld bytes) which is greater than total scratch space (%ld bytes)\nIncrease size of collective scratch space through GASNET_COLL_SCRATCH_SIZE environment variable to at least %ld bytes\n", 
                        scratch_req->team->myrank, 
-                       (size_t)scratch_req->incoming_size, (size_t)scratch_req->team->scratch_segs[0].size, (size_t)scratch_req->incoming_size); 
+                       (long int)scratch_req->incoming_size, (long int)scratch_req->team->scratch_segs[0].size, (long int)scratch_req->incoming_size); 
   }
   gasneti_assert(op->waiting_scratch_op == 0 || op->waiting_scratch_op == 1);
   if(op->waiting_scratch_op) {
